requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
alpha_vantage==3.0.0
amqp==5.3.1
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
autobahn==19.11.2
Automat==24.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==43.0.3
dhanhq==2.0.1
Django==5.1.3
django-celery-beat==2.7.0
django-extensions==3.2.3
django-timezone-field==7.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
exceptiongroup==1.2.2
Flask==3.1.0
frozendict==2.4.6
frozenlist==1.5.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gspread==6.1.4
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
kiteconnect==5.0.1
kombu==5.4.2
logzero==1.7.0
lxml==5.3.0
MarkupSafe==3.0.2
multidict==6.1.0
multitasking==0.0.11
numpy==2.1.3
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
peewee==3.17.8
platformdirs==4.3.6
plyer==2.1.0
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pyOpenSSL==24.2.1
pyotp==2.8.0
pyparsing==3.2.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.8
pytz==2024.2
pywin32==305
rdflib==6.3.2
rdflib-jsonld==0.6.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
service-identity==24.2.0
simplejson==3.19.1
six==1.16.0
smartapi-python==1.5.2
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.2
ta==0.11.0
tomli==2.2.1
Twisted==24.11.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
xlwings==0.33.4
yarl==1.18.3
yfinance==0.2.50
zope.interface==7.2
